changed routing api/v1/vt/livestream -> api/v1/vehicles/telematics/livestream api/v1/vt/vod -> api/v1/vehicles/video api/v1/vt/lookups -> api/v1/vehicles/video/lookups
