+++ config/routes.rb
@@ -4,11 +4,14 @@ Rails.application.routes.draw do
       post "vehicle/telematics/last_known", to: "vehicle_stats#last_known"
       resources :drivers, only: [:index]
 
-      # VT APIs
-      namespace :vt do
-        resources :livestream, only: %i[index create update destroy], controller: "livestream"
-        resources :vod, only: %i[index create], controller: "vod"
-        resources :lookups, only: [:index], controller: "data"
+      namespace :vehicle do
+        resources :video, only: %i[index create], controller: "vod"
+        namespace :telematics do
+          resources :livestream, only: %i[index create update destroy], controller: "livestream"
+        end
+        namespace :video do
+          resources :lookups, only: [:index], controller: "data"
+        end
       end
     end
   end
